add pagination

+++ routes/web.php
@@ -19,7 +19,7 @@ Route::get('/pinjam/create', [BookController::class, 'create']);
 Route::get('/pinjam', [BookController::class, 'index']);
 Route::get('/pinjam/{id}', [BookController::class, 'show']);
 Route::post('/pinjam', [BookController::class, 'store']);
-Route::get('/pinjam/{id}/edit', [BookController::class, 'edit']);
-Route::patch('/pinjam/{id}', [BookController::class, 'update']);
-Route::delete('/pinjam/{id}', [BookController::class, 'delete']);
+Route::get('/pinjam/{id}/edit', [BookController::class, 'edit'])->middleware('is_admin');
+Route::patch('/pinjam/{id}', [BookController::class, 'update'])->middleware('is_admin');
+Route::delete('/pinjam/{id}', [BookController::class, 'delete'])->middleware('is_admin');
